requirements.txt
pandas
numpy
jupyter
matplotlib
mplfinance
Riskfolio-Lib
norgatedata
zipline-reloaded
zipline-norgatedata
yfinance
quantstats
